services/auth-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.1
pydantic-settings>=2.1.0
structlog>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
prometheus-fastapi-instrumentator>=6.1.0
aiosmtplib>=3.0.0
jinja2>=3.1.2
pyotp>=2.9.0
qrcode[pil]>=7.4.2
slowapi>=0.1.9
zxcvbn-python>=0.5.0
arq>=0.25.0